Implements usePolling hook, implements stale while revalidate on usePaginatedRecords
